Startup fixes

+++ markov.py
@@ -139,30 +139,36 @@ class Markov(CogHelper):
     '''
     def __init__(self, bot, db_engine, logger, settings):
         super().__init__(bot, db_engine, logger, settings)
-        BASE.metadata.create_all(self.db_engine)
-        BASE.metadata.bind = self.db_engine
+
+        self._task = None
+        self.enabled = True
         try:
             validate_config(settings['markov'], MARKOV_SECTION_SCHEMA)
         except ValidationError as exc:
             raise CogMissingRequiredArg('Invalid config given for markov bot') from exc
         except KeyError:
             settings['markov'] = {}
+            self.enabled = False
+            return
+        BASE.metadata.create_all(self.db_engine)
+        BASE.metadata.bind = self.db_engine
         self.loop_sleep_interval = settings['markov'].get('loop_sleep_interval', LOOP_SLEEP_INTERVAL_DEFAULT)
         self.message_check_limit = settings['markov'].get('message_check_limit', MESSAGE_CHECK_LIMIT)
         self.history_retention_days = settings['markov'].get('history_retention_days', MARKOV_HISTORY_RETENTION_DAYS_DEFAULT)
         self.server_reject_list = settings['markov'].get('server_reject_list', [])
 
         self.lock_file = Path(NamedTemporaryFile(delete=False).name) #pylint:disable=consider-using-with
-        self._task = None
 
     async def cog_load(self):
-        self._task = self.bot.loop.create_task(self.main_loop())
+        if self.enabled:
+            self._task = self.bot.loop.create_task(self.main_loop())
 
     async def cog_unload(self):
-        if self._task:
-            self._task.cancel()
-        if self.lock_file.exists():
-            self.lock_file.unlink()
+        if self.enabled:
+            if self._task:
+                self._task.cancel()
+            if self.lock_file.exists():
+                self.lock_file.unlink()
 
     def __ensure_word(self, word):
         if len(word) >= MAX_WORD_LENGTH:
